comparison/requirements.txt
gentis-ai
crewai
google-generativeai
python-dotenv
langchain-google-genai
